(PC-22884)[PRO] refactor: Refactored `ico-pen-black` usage.
